feat(Input): move styles to Base theme
