+++ docs/src/examples/components/Input/Types/InputExample.shorthand.steps.ts
@@ -2,6 +2,7 @@ import { Input } from '@stardust-ui/react'
 
 const config: ScreenerTestsConfig = {
   steps: [builder => builder.focus(`.${Input.className} input`).snapshot('Can be focused')],
+  themes: ['base', 'teams', 'teamsDark', 'teamsHighContrast'],
 }
 
 export default config
