+++ packages/react/src/themes/base/componentStyles.ts
@@ -5,6 +5,8 @@ export { default as FlexItem } from './components/Flex/flexItemStyles'
 
 export { default as Icon } from './components/Icon/iconStyles'
 
+export { default as Input } from './components/Input/inputStyles'
+
 export { default as Loader } from './components/Loader/loaderStyles'
 
 export { default as ProviderBox } from './components/Provider/providerBoxStyles'
