+++ packages/react/src/themes/base/componentVariables.ts
@@ -5,6 +5,8 @@ export { default as FlexItem } from './components/Flex/flexItemVariables'
 
 export { default as Icon } from './components/Icon/iconVariables'
 
+export { default as Input } from './components/Input/inputVariables'
+
 export { default as Loader } from './components/Loader/loaderVariables'
 
 export { default as ProviderBox } from './components/Provider/providerBoxVariables'
